Implement detect_feature_types function
detect_feature_types detects the features for each feature column.
The feature types it detects is numerical and categorical. Then stores
it in a list of data type Feature, Feature class is an wrapper that
stores the name of the data column and the feature type of the data
column. Then returns this list of Feature objects.

Daan Rosies (S5631742)
15:55 | 10/17/2024