Implemented the feature class
Implemented the feature class. Which has two properties for now
the type property which defines which type of feature column it is
numerical or catigorical. The name property which is the name
of the feature column. Furthermore, type can only take two values
with are numerical or categorical any other value will raise a valueError.

The class still needs to be edited for style conventions, docstrings
and other implementations. As well as, best case OOP practises.

Daan Rosies (S5631742)